Split resolve/media into pool adapter and media operations
resolve/media.py fused two halves: the media pool adapter every other
package consumed (clip location, property reading, bin navigation, frame
bounds, offline and still handling) and the six tool-facing operations
that are themselves callers of it.

The adapter moves to resolve/pool.py; resolve/media.py keeps import,
list, inspect, metadata, organize and relink. External callers now import
the adapter directly as `mediapool` rather than reaching pool names
through the operations module - mypy strict's no-implicit-reexport rule
holds that line. Three helpers changed shape for the seam: _clips_under
and _looks_like_image become public (clips_under, looks_like_image),
_number gains a named public form (frame_count), and _wants_recursion
moves to the operations side, where both its callers live.

The media test file splits the same way: test_media_pool.py takes the 28
bin-addressing, lookup, offline and frame-bounds tests, test_media_tools.py
keeps the 43 that are about what an operation does. All 71 survive.

Refs #227
